requirements.txt
jsonschema==4.19.0
jinja2==3.1.2
requests==2.31.0
python-dotenv==1.0.0
openai==1.32.0
openai>=1.0.0
python-dotenv>=1.0.0
jsonschema>=4.0.0
